backend/requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
more-itertools==10.6.0
pdfkit==1.0.0
python-dotenv==1.1.0
razorpay==1.4.2
requests==2.32.3
setuptools==78.1.0
SQLAlchemy==2.0.40
typeguard==4.4.2
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
